include extensions to struct
