+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod create_structs;
+pub mod element_utils;
 pub mod string_utils;
