+++ structs.rs
@@ -1,40 +1,74 @@
 
-pub struct FinancingActApplicationGeometriesType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "FinancingActApplicationGeometry")]
-    pub financing_act_application_geometry: Vec<FinancingActGeometryType>,
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActGeometryType {
+    #[serde(rename = "@id")]
+    pub id: IdStringType,
+    #[serde(rename = "FeatureCode")]
+    pub feature_code: FeatureCodeType,
+    #[serde(rename = "FeatureInfo")]
+    pub feature_info: String500Type,
+    #[serde(rename = "PointAndLineGeometriesGroup")]
+    pub point_and_line_geometries_group: String,
 }
 
-pub struct WorkingRepresentativeType {
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SubsidyApplierReferenceListType {
+    #[serde(rename = "SubsidyApplierReference")]
+    pub subsidy_applier_reference: Vec<SubsidyApplierReferenceType>,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
 pub struct FinancingActApplicationStandsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
     #[serde(rename = "FinancingActApplicationStand")]
     pub financing_act_application_stand: Vec<FinancingActApplicationStandType>,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CostTypeAndCompletedWorkApplicationRealEstateType {
+    #[serde(rename = "RealEstateId")]
+    pub real_estate_id: IdStringNotEmptyType,
+    #[serde(rename = "CostType")]
+    pub cost_type: CostTypeType,
+    #[serde(rename = "PlannedWorkAmount", skip_serializing_if = "Option::is_none")]
+    pub planned_work_amount: Option<AmountType>,
+    #[serde(rename = "UnitPrice", skip_serializing_if = "Option::is_none")]
+    pub unit_price: Option<UnitPriceType>,
+    #[serde(rename = "SubsidyAmount")]
+    pub subsidy_amount: MoneyType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
 pub struct FinancingActApplicationOtherSubjectsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
     #[serde(rename = "CostTypeAndCompletedWorkApplicationRealEstate")]
     pub cost_type_and_completed_work_application_real_estate: CostTypeAndCompletedWorkApplicationRealEstateType,
     #[serde(rename = "FinancingActApplicationGeometries")]
     pub financing_act_application_geometries: FinancingActApplicationGeometriesType,
 }
 
-pub struct SubsidyApplierReferenceListType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "SubsidyApplierReference")]
-    pub subsidy_applier_reference: Vec<SubsidyApplierReferenceType>,
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActCompletionGeometriesType {
+    #[serde(rename = "FinancingActCompletionGeometry")]
+    pub financing_act_completion_geometry: Vec<FinancingActGeometryType>,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
+pub struct OwnerShipTypeType {
+    #[serde(rename = "OwnerShipTypeCode")]
+    pub owner_ship_type_code: OwnerShipTypeCodeType,
+    #[serde(rename = "AdditionalInfo", skip_serializing_if = "Option::is_none")]
+    pub additional_info: Option<String500Type>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActCompletionOtherSubjectsType {
+    #[serde(rename = "CostTypeAndCompletedWorkCompletionRealEstate")]
+    pub cost_type_and_completed_work_completion_real_estate: CostTypeAndCompletedWorkCompletionRealEstateType,
+    #[serde(rename = "FinancingActCompletionGeometries")]
+    pub financing_act_completion_geometries: FinancingActCompletionGeometriesType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
 pub struct CostTypeAndCompletedWorkCompletionRealEstateType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
     #[serde(rename = "RealEstateId")]
     pub real_estate_id: IdStringNotEmptyType,
     #[serde(rename = "CostType")]
@@ -49,9 +83,188 @@ pub struct CostTypeAndCompletedWorkCompletionRealEstateType {
     pub subsidy_amount: MoneyType,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ForestCentreDataType {
+    #[serde(rename = "@id")]
+    pub id: IdStringType,
+    #[serde(rename = "@schemaPackageVersion")]
+    pub schema_package_version: ForestDataStandardSchemaPackageVersionType,
+    #[serde(rename = "@schemaPackageSubversion")]
+    pub schema_package_subversion: ForestDataStandardSchemaPackageSubversionType,
+    #[serde(rename = "@schemaPackageVersionDate")]
+    pub schema_package_version_date: DateType,
+    #[serde(rename = "TimeStamp")]
+    pub time_stamp: TimeStampType,
+    #[serde(rename = "Message")]
+    pub message: ForestCentreDataMessageType,
+    #[serde(rename = "RequestReference", skip_serializing_if = "Option::is_none")]
+    pub request_reference: Option<ReferenceType>,
+    #[serde(rename = "MetadataText", skip_serializing_if = "Option::is_none")]
+    pub metadata_text: Option<String1000Type>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct PayeeType {
+    #[serde(flatten)]
+    pub base: ContactInformationType,
+    #[serde(rename = "BankAccount", skip_serializing_if = "Option::is_none")]
+    pub bank_account: Option<BankAccountType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ApplicationActorType {
+    #[serde(flatten)]
+    pub base: ContactInformationType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActCompletionStandsType {
+    #[serde(rename = "FinancingActCompletionStand")]
+    pub financing_act_completion_stand: Vec<FinancingActCompletionStandType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct PayeesAndRealEstatesType {
+    #[serde(rename = "PayeeAndRealEstate")]
+    pub payee_and_real_estate: Vec<PayeeAndRealEstateType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct DocumentsType {
+    #[serde(rename = "Document")]
+    pub document: Vec<Document4MBType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActApplicationGeometriesType {
+    #[serde(rename = "FinancingActApplicationGeometry")]
+    pub financing_act_application_geometry: Vec<FinancingActGeometryType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActApplicationStandType {
+    #[serde(rename = "@id")]
+    pub id: IdStringType,
+    #[serde(rename = "StandNumber")]
+    pub stand_number: StandNumberType,
+    #[serde(rename = "StandNumberExtension", skip_serializing_if = "Option::is_none")]
+    pub stand_number_extension: Option<StandNumberExtensionType>,
+    #[serde(rename = "StandReference")]
+    pub stand_reference: ReferenceType,
+    #[serde(rename = "Area")]
+    pub area: AreaType,
+    #[serde(rename = "MainGroup", skip_serializing_if = "Option::is_none")]
+    pub main_group: Option<ExtendedMainGroupType>,
+    #[serde(rename = "SubGroup", skip_serializing_if = "Option::is_none")]
+    pub sub_group: Option<SubGroupType>,
+    #[serde(rename = "FertilityClass", skip_serializing_if = "Option::is_none")]
+    pub fertility_class: Option<FertilityClassType>,
+    #[serde(rename = "SoilType", skip_serializing_if = "Option::is_none")]
+    pub soil_type: Option<SoilTypeType>,
+    #[serde(rename = "MeanHeight", skip_serializing_if = "Option::is_none")]
+    pub mean_height: Option<HeightType>,
+    #[serde(rename = "DiameterClass", skip_serializing_if = "Option::is_none")]
+    pub diameter_class: Option<DiameterClassType>,
+    #[serde(rename = "CostTypeAndCompletedWorkApplication")]
+    pub cost_type_and_completed_work_application: CostTypeAndCompletedWorkApplicationType,
+    #[serde(rename = "StandExtraInfo", skip_serializing_if = "Option::is_none")]
+    pub stand_extra_info: Option<String2000Type>,
+    #[serde(rename = "polygonProperty")]
+    pub polygon_property: String,
+    #[serde(rename = "LocationEstate")]
+    pub location_estate: LocationEstateType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct AttorneyType {
+    #[serde(flatten)]
+    pub base: ContactInformationType,
+    #[serde(rename = "PowerOfAttorney")]
+    pub power_of_attorney: PowerOfAttorneyType,
+    #[serde(rename = "RightToSpecifyBankAccountsOfPaymentTransactions")]
+    pub right_to_specify_bank_accounts_of_payment_transactions: YesNoType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActRealEstatesType {
+    #[serde(rename = "FinancingActRealEstate")]
+    pub financing_act_real_estate: Vec<FinancingActRealEstateType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ApplicationActorsType {
+    #[serde(rename = "ApplicationActor")]
+    pub application_actor: Vec<ApplicationActorType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CompletionDeclarationActorsType {
+    #[serde(rename = "CompletionDeclarationActor")]
+    pub completion_declaration_actor: Vec<PayeeType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CostTypeType {
+    #[serde(rename = "CostTypeNumber")]
+    pub cost_type_number: CostTypeNumberType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct WorkingRepresentativesType {
+    #[serde(rename = "WorkingRepresentative")]
+    pub working_representative: Vec<WorkingRepresentativeType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SubsidyApplierReferenceType {
+    #[serde(rename = "SubsidyApplierId")]
+    pub subsidy_applier_id: IdStringNotEmptyType,
+    #[serde(rename = "MainApplier")]
+    pub main_applier: YesNoType,
+    #[serde(rename = "ShareOfOwnerShip", skip_serializing_if = "Option::is_none")]
+    pub share_of_owner_ship: Option<PercentType>,
+    #[serde(rename = "OwnerShipType")]
+    pub owner_ship_type: OwnerShipTypeType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SubsidyAppliersType {
+    #[serde(rename = "SubsidyApplier")]
+    pub subsidy_applier: Vec<PayeeType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CostTypeAndCompletedWorkCompletionType {
+    #[serde(rename = "CostType")]
+    pub cost_type: CostTypeType,
+    #[serde(rename = "CompletedWorkAmount")]
+    pub completed_work_amount: AmountType,
+    #[serde(rename = "UnitPrice", skip_serializing_if = "Option::is_none")]
+    pub unit_price: Option<UnitPriceType>,
+    #[serde(rename = "SubsidyAmount")]
+    pub subsidy_amount: MoneyType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct PayeeAndRealEstateType {
+    #[serde(rename = "RealEstateId")]
+    pub real_estate_id: IdStringNotEmptyType,
+    #[serde(rename = "PayeeId")]
+    pub payee_id: IdStringNotEmptyType,
+    #[serde(rename = "ParticipationPercentage", skip_serializing_if = "Option::is_none")]
+    pub participation_percentage: Option<PercentType>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct FinancingActRealEstateType {
+    #[serde(flatten)]
+    pub base: BaseRealEstateType2,
+    #[serde(rename = "SubsidyApplierReferenceList")]
+    pub subsidy_applier_reference_list: SubsidyApplierReferenceListType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
 pub struct AuthorizationOfAttorneyType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
     #[serde(rename = "Authorization")]
     pub authorization: YesNoType,
     #[serde(rename = "Date")]
@@ -64,11 +277,28 @@ pub struct AuthorizationOfAttorneyType {
     pub power_of_attorney_document: Option<base64Binary>,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
+pub struct WorkingRepresentativeType {
+    #[serde(flatten)]
+    pub base: ContactInformationType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CostTypeAndCompletedWorkApplicationType {
+    #[serde(rename = "CostType")]
+    pub cost_type: CostTypeType,
+    #[serde(rename = "PlannedWorkAmount")]
+    pub planned_work_amount: AmountType,
+    #[serde(rename = "UnitPrice", skip_serializing_if = "Option::is_none")]
+    pub unit_price: Option<UnitPriceType>,
+    #[serde(rename = "SubsidyAmount")]
+    pub subsidy_amount: MoneyType,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
 pub struct FinancingActCompletionStandType {
     #[serde(rename = "@id")]
     pub id: IdStringType,
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
     #[serde(rename = "StandNumber")]
     pub stand_number: StandNumberType,
     #[serde(rename = "StandNumberExtension", skip_serializing_if = "Option::is_none")]
@@ -113,252 +343,4 @@ pub struct FinancingActCompletionStandType {
     pub polygon_property: String,
     #[serde(rename = "LocationEstate")]
     pub location_estate: LocationEstateType,
-}
-
-pub struct ForestCentreDataType {
-    #[serde(rename = "@id")]
-    pub id: IdStringType,
-    #[serde(rename = "@schemaPackageVersion")]
-    pub schema_package_version: ForestDataStandardSchemaPackageVersionType,
-    #[serde(rename = "@schemaPackageSubversion")]
-    pub schema_package_subversion: ForestDataStandardSchemaPackageSubversionType,
-    #[serde(rename = "@schemaPackageVersionDate")]
-    pub schema_package_version_date: DateType,
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "TimeStamp")]
-    pub time_stamp: TimeStampType,
-    #[serde(rename = "Message")]
-    pub message: ForestCentreDataMessageType,
-    #[serde(rename = "RequestReference", skip_serializing_if = "Option::is_none")]
-    pub request_reference: Option<ReferenceType>,
-    #[serde(rename = "MetadataText", skip_serializing_if = "Option::is_none")]
-    pub metadata_text: Option<String1000Type>,
-}
-
-pub struct DocumentsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "Document")]
-    pub document: Vec<Document4MBType>,
-}
-
-pub struct CostTypeAndCompletedWorkApplicationRealEstateType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "RealEstateId")]
-    pub real_estate_id: IdStringNotEmptyType,
-    #[serde(rename = "CostType")]
-    pub cost_type: CostTypeType,
-    #[serde(rename = "PlannedWorkAmount", skip_serializing_if = "Option::is_none")]
-    pub planned_work_amount: Option<AmountType>,
-    #[serde(rename = "UnitPrice", skip_serializing_if = "Option::is_none")]
-    pub unit_price: Option<UnitPriceType>,
-    #[serde(rename = "SubsidyAmount")]
-    pub subsidy_amount: MoneyType,
-}
-
-pub struct CostTypeType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "CostTypeNumber")]
-    pub cost_type_number: CostTypeNumberType,
-}
-
-pub struct WorkingRepresentativesType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "WorkingRepresentative")]
-    pub working_representative: Vec<WorkingRepresentativeType>,
-}
-
-pub struct CostTypeAndCompletedWorkApplicationType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "CostType")]
-    pub cost_type: CostTypeType,
-    #[serde(rename = "PlannedWorkAmount")]
-    pub planned_work_amount: AmountType,
-    #[serde(rename = "UnitPrice", skip_serializing_if = "Option::is_none")]
-    pub unit_price: Option<UnitPriceType>,
-    #[serde(rename = "SubsidyAmount")]
-    pub subsidy_amount: MoneyType,
-}
-
-pub struct CostTypeAndCompletedWorkCompletionType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "CostType")]
-    pub cost_type: CostTypeType,
-    #[serde(rename = "CompletedWorkAmount")]
-    pub completed_work_amount: AmountType,
-    #[serde(rename = "UnitPrice", skip_serializing_if = "Option::is_none")]
-    pub unit_price: Option<UnitPriceType>,
-    #[serde(rename = "SubsidyAmount")]
-    pub subsidy_amount: MoneyType,
-}
-
-pub struct ApplicationActorType {
-}
-
-pub struct FinancingActCompletionStandsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "FinancingActCompletionStand")]
-    pub financing_act_completion_stand: Vec<FinancingActCompletionStandType>,
-}
-
-pub struct CompletionDeclarationActorsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "CompletionDeclarationActor")]
-    pub completion_declaration_actor: Vec<PayeeType>,
-}
-
-pub struct FinancingActCompletionGeometriesType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "FinancingActCompletionGeometry")]
-    pub financing_act_completion_geometry: Vec<FinancingActGeometryType>,
-}
-
-pub struct PayeeType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "BankAccount", skip_serializing_if = "Option::is_none")]
-    pub bank_account: Option<BankAccountType>,
-}
-
-pub struct AttorneyType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "PowerOfAttorney")]
-    pub power_of_attorney: PowerOfAttorneyType,
-    #[serde(rename = "RightToSpecifyBankAccountsOfPaymentTransactions")]
-    pub right_to_specify_bank_accounts_of_payment_transactions: YesNoType,
-}
-
-pub struct FinancingActRealEstatesType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "FinancingActRealEstate")]
-    pub financing_act_real_estate: Vec<FinancingActRealEstateType>,
-}
-
-pub struct PayeesAndRealEstatesType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "PayeeAndRealEstate")]
-    pub payee_and_real_estate: Vec<PayeeAndRealEstateType>,
-}
-
-pub struct OwnerShipTypeType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "OwnerShipTypeCode")]
-    pub owner_ship_type_code: OwnerShipTypeCodeType,
-    #[serde(rename = "AdditionalInfo", skip_serializing_if = "Option::is_none")]
-    pub additional_info: Option<String500Type>,
-}
-
-pub struct ApplicationActorsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "ApplicationActor")]
-    pub application_actor: Vec<ApplicationActorType>,
-}
-
-pub struct FinancingActApplicationStandType {
-    #[serde(rename = "@id")]
-    pub id: IdStringType,
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "StandNumber")]
-    pub stand_number: StandNumberType,
-    #[serde(rename = "StandNumberExtension", skip_serializing_if = "Option::is_none")]
-    pub stand_number_extension: Option<StandNumberExtensionType>,
-    #[serde(rename = "StandReference")]
-    pub stand_reference: ReferenceType,
-    #[serde(rename = "Area")]
-    pub area: AreaType,
-    #[serde(rename = "MainGroup", skip_serializing_if = "Option::is_none")]
-    pub main_group: Option<ExtendedMainGroupType>,
-    #[serde(rename = "SubGroup", skip_serializing_if = "Option::is_none")]
-    pub sub_group: Option<SubGroupType>,
-    #[serde(rename = "FertilityClass", skip_serializing_if = "Option::is_none")]
-    pub fertility_class: Option<FertilityClassType>,
-    #[serde(rename = "SoilType", skip_serializing_if = "Option::is_none")]
-    pub soil_type: Option<SoilTypeType>,
-    #[serde(rename = "MeanHeight", skip_serializing_if = "Option::is_none")]
-    pub mean_height: Option<HeightType>,
-    #[serde(rename = "DiameterClass", skip_serializing_if = "Option::is_none")]
-    pub diameter_class: Option<DiameterClassType>,
-    #[serde(rename = "CostTypeAndCompletedWorkApplication")]
-    pub cost_type_and_completed_work_application: CostTypeAndCompletedWorkApplicationType,
-    #[serde(rename = "StandExtraInfo", skip_serializing_if = "Option::is_none")]
-    pub stand_extra_info: Option<String2000Type>,
-    #[serde(rename = "polygonProperty")]
-    pub polygon_property: String,
-    #[serde(rename = "LocationEstate")]
-    pub location_estate: LocationEstateType,
-}
-
-pub struct SubsidyAppliersType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "SubsidyApplier")]
-    pub subsidy_applier: Vec<PayeeType>,
-}
-
-pub struct SubsidyApplierReferenceType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "SubsidyApplierId")]
-    pub subsidy_applier_id: IdStringNotEmptyType,
-    #[serde(rename = "MainApplier")]
-    pub main_applier: YesNoType,
-    #[serde(rename = "ShareOfOwnerShip", skip_serializing_if = "Option::is_none")]
-    pub share_of_owner_ship: Option<PercentType>,
-    #[serde(rename = "OwnerShipType")]
-    pub owner_ship_type: OwnerShipTypeType,
-}
-
-pub struct FinancingActRealEstateType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "SubsidyApplierReferenceList")]
-    pub subsidy_applier_reference_list: SubsidyApplierReferenceListType,
-}
-
-pub struct FinancingActGeometryType {
-    #[serde(rename = "@id")]
-    pub id: IdStringType,
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "FeatureCode")]
-    pub feature_code: FeatureCodeType,
-    #[serde(rename = "FeatureInfo")]
-    pub feature_info: String500Type,
-    #[serde(rename = "PointAndLineGeometriesGroup")]
-    pub point_and_line_geometries_group: String,
-}
-
-pub struct FinancingActCompletionOtherSubjectsType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "CostTypeAndCompletedWorkCompletionRealEstate")]
-    pub cost_type_and_completed_work_completion_real_estate: CostTypeAndCompletedWorkCompletionRealEstateType,
-    #[serde(rename = "FinancingActCompletionGeometries")]
-    pub financing_act_completion_geometries: FinancingActCompletionGeometriesType,
-}
-
-pub struct PayeeAndRealEstateType {
-    #[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(rename = "RealEstateId")]
-    pub real_estate_id: IdStringNotEmptyType,
-    #[serde(rename = "PayeeId")]
-    pub payee_id: IdStringNotEmptyType,
-    #[serde(rename = "ParticipationPercentage", skip_serializing_if = "Option::is_none")]
-    pub participation_percentage: Option<PercentType>,
 }
